LayoutTests/platform/glib/http/tests/websocket/tests/hybi/interleaved-fragments-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/interleaved-fragments' failed: Received invalid WebSocket response from the server
Test whether WebSocket rejects interleaved fragmented frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
